SRSs: code cleanup, introduced SRSCache
