+++ src/geo/points.test.ts
@@ -1,10 +1,7 @@
-import { SRSs } from '../srs'
 import { PointFactory } from './pointFactory'
 import { Points } from './points'
 
 describe('Points test', () => {
-  beforeAll(SRSs.init)
-
   test('parsing incomplete coordinate (missing srs)', () => {
     const parsed = Points.parse('POINT(144.50234 -6.321367)')
     expect(parsed).toBeNull()
