+++ src/srs/srss.ts
@@ -1,36 +1,7 @@
 import { SRS } from './srs'
-import { SRSFactory } from './factory'
+import { SRSCache } from './srsCache'
 
-const formatName = (name = ''): string => name.replace(/_/g, ' ')
-
-let SRS_MAP: { [code: string]: SRS }
-let SRS_ARRAY: SRS[]
-
-const init = async (): Promise<void> => {
-  if (!SRS_MAP) {
-    SRS_MAP = {}
-    const addCrs = (crs: { name: string; wkid: number; wkt: string }): void => {
-      const { name, wkid, wkt } = crs
-      const code: string = wkid.toString()
-      SRS_MAP[code] = SRSFactory.createInstance({ code, name: formatName(name), wkt })
-    }
-
-    const [{ GeographicCoordinateSystems }, { ProjectedCoordinateSystems }] = await Promise.all([
-      import('@esri/proj-codes/pe_list_geogcs.json'),
-      import('@esri/proj-codes/pe_list_projcs.json'),
-    ])
-    GeographicCoordinateSystems.forEach(addCrs)
-    ProjectedCoordinateSystems.forEach(addCrs)
-
-    SRS_ARRAY = Object.values(SRS_MAP).sort((srs1, srs2) => srs1.name.localeCompare(srs2.name))
-  }
-}
-
-const getSRSByCode = async (code: string): Promise<SRS | undefined> => {
-  await init()
-
-  return SRS_MAP[code]
-}
+const getSRSByCode = async (code: string): Promise<SRS | undefined> => (await SRSCache.getMap())[code]
 
 /**
  * Finds a list of srs whose name or code matches the specified parameter.
@@ -40,18 +11,20 @@ const getSRSByCode = async (code: string): Promise<SRS | undefined> => {
  * @returns {Promise<SRS[]>} - List of SRS matching the specified code or name.
  */
 const findSRSByCodeOrName = async (codeOrName: string, limit = 200): Promise<SRS[]> => {
-  await init()
+  const srsArray = await SRSCache.getArray()
 
   const contains = (string: string, value: string) => string.indexOf(value) >= 0
   const codeOrNameLowerCase = codeOrName.toLocaleLowerCase()
-  return SRS_ARRAY.filter(
-    (srs) =>
-      contains(srs.code.toLowerCase(), codeOrNameLowerCase) || contains(srs.name.toLowerCase(), codeOrNameLowerCase)
-  ).slice(0, limit)
+
+  return srsArray
+    .filter(
+      (srs) =>
+        contains(srs.code.toLowerCase(), codeOrNameLowerCase) || contains(srs.name.toLowerCase(), codeOrNameLowerCase)
+    )
+    .slice(0, limit)
 }
 
 export const SRSs = {
   findSRSByCodeOrName,
   getSRSByCode,
-  init,
 }
